Reformatting, reload page after deleting from board.

+++ src/main/java/com/gbdevteam/teamnotes/controller/BoardController.java
@@ -3,14 +3,19 @@ package com.gbdevteam.teamnotes.controller;
 import com.gbdevteam.teamnotes.controller.validators.ValidUUID;
 import com.gbdevteam.teamnotes.dto.BoardDTO;
 import com.gbdevteam.teamnotes.dto.NoteDTO;
+import com.gbdevteam.teamnotes.model.BoardRoleEnum;
+import com.gbdevteam.teamnotes.service.BoardRoleService;
 import com.gbdevteam.teamnotes.service.BoardService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.nio.file.attribute.UserPrincipal;
+import java.security.Principal;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
@@ -22,6 +27,7 @@ import java.util.UUID;
 @Validated
 public class BoardController {
     private final BoardService boardService;
+    private final BoardRoleService boardRoleService;
 
     @GetMapping("/user/{userId}")
     public List<BoardDTO> findAll(@PathVariable("userId") @ValidUUID UUID userId) {
@@ -45,14 +51,20 @@ public class BoardController {
     }
 
     @PutMapping
-    public void update(@Valid @RequestBody BoardDTO board) {
-        boardService.update(board);
+    public void update(@Valid @RequestBody BoardDTO board, Principal principal) {
+//        BoardRoleEnum boardRole = boardRoleService.checkRole(board.getId(), principal.getName());
+//        if(boardRole.equals(BoardRoleEnum.OWNER)) {
+            boardService.update(board);
+//        }
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteById(@PathVariable("id") UUID id) {
-        boardService.deleteById(id);
+    public void deleteById(@PathVariable("id") UUID id, Principal principal) {
+        BoardRoleEnum boardRole = boardRoleService.checkRole(id, principal.getName());
+        if(boardRole.equals(BoardRoleEnum.OWNER)) {
+            boardService.deleteById(id);
+        }
     }
 
     @GetMapping("/{boardId}/addUser/{userId}")
